hcp-crm-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
langchain-core
langgraph
langchain-groq
python-dotenv
pymysql
cryptography